EES-2415 tidying and testing
