+++ src/explore-education-statistics-common/src/services/publicationService.ts
@@ -53,6 +53,14 @@ export interface PublicationSummary {
   summary: string;
 }
 
+export interface BasicPublicationContact {
+  id?: string;
+  contactName: string;
+  contactTelNo: string;
+  teamEmail: string;
+  teamName?: string;
+}
+
 export interface PublicationContact {
   teamName: string;
   teamEmail: string;
